requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pymysql==1.1.0
cryptography==42.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
boto3==1.34.34
PyPDF2==3.0.1
reportlab==4.0.9
httpx==0.26.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0
redis==5.0.1
slowapi==0.1.9
geopy==2.4.1
